Rework documentation URLs as inline markdown
Redesign documentations with markdown text

Redesign way to document scripts/categories and present the
documentation.

Documentation is show in an expandable box instead of tooltip. This is
to allow writing longer documentation (tooltips are meant to be used for
short text) and have better experience on mobile.

If a node (script/category) has documentation it's now shown with single
information icon (ℹ) aligned to right.

Add support for rendering documentation as markdown.